Fix/2023/point (#27)
* use i64/u64 in point instead of isize/usize

* day 3

* day 10

* day 11

* day 16

* day 16

* day 17

* fix 14

* day 10p1

+++ 2023/03/src/part1.rs
@@ -15,7 +15,7 @@ fn get_numbers(lines: Vec<String>) -> Vec<u64> {
     let mut res: Vec<u64> = Vec::new();
 
     for i in 0..lines.len() {
-        let x = i as isize;
+        let x = i as i64;
         let mut num: Option<u64> = None;
         let mut valid = false;
         let line = lines[i].as_bytes();
@@ -26,7 +26,7 @@ fn get_numbers(lines: Vec<String>) -> Vec<u64> {
                 num = Some(num.unwrap_or_default() * 10 + ((c - b'0') as u64));
                 if !valid {
                     // calc once
-                    valid = Point { x, y: j as isize }
+                    valid = Point { x, y: j as i64 }
                         .neighbors()
                         .iter()
                         .any(|p| symbols.contains(p));
@@ -58,7 +58,7 @@ fn get_symbols(lines: &Vec<String>) -> HashSet<Point> {
     let mut res: HashSet<Point> = HashSet::new();
 
     for i in 0..lines.len() {
-        let x = i as isize;
+        let x = i as i64;
         let line = lines[i].as_bytes();
         for j in 0..line.len() {
             let c = line[j];
@@ -67,7 +67,7 @@ fn get_symbols(lines: &Vec<String>) -> HashSet<Point> {
                 continue;
             }
 
-            res.insert(Point { x, y: j as isize });
+            res.insert(Point { x, y: j as i64 });
         }
     }
